requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==38.0.4
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
Flask==2.2.5
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.1
httpq==1.2.0
httptools==0.5.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.2
mitm==1.5.0
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
msgpack==1.0.5
multidict==6.0.4
orjson==3.8.12
passlib==1.7.4
pbr==5.9.0
protobuf==4.23.1
proxy.py==2.4.3
publicsuffix2==2.20191221
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
toolbox==1.11.0
tornado==6.3.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==2.0.2
urwid==2.1.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
zstandard==0.19.0
